requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SpeechRecognition==3.8.1
uritemplate==3.0.1
urllib3==1.25.10
